Routes changes
(change) WorkshopOfficeWorkList: The employee will not be considered anymore due to the DB change and also we will need the workshop office id to get the employee list in the view. Also, we will use left outer join to get more than one workshop office work in the list and we will group that list by the id, so we don't get repeated rows.
(new) AddWorkshopOfficeWork: To add a workshop office work with its correspondent milestones when the user confirms that they want to receive the service.