Changes in the following fields:
payment receipt suscription: Shouldn't make the FK to PK
usability questionnaire: We need the status as well
Question item: The ID needs to be autoincrementable
User Usability questionnaire:  The ID needs to be autoincrementable
Work case: ID needs to be autoincrementable. Also, we need the date time when the case was created.